+++ src/gui/GUIElements/Clickable.ts
@@ -1,0 +1,7 @@
+export interface Clickable {
+  onClick(): void;
+  onHover(): void;
+  isCoordInElement(x: number, y: number): boolean;
+  width: number;
+  height: number;
+}
